+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## UNRELEASED
+
+* Issue-291: Update `earthacess` version to enable use of Earthdata token for
+  authentication.
+
 ## v1.14.0rc0 (2026-06-05)
 
 * PSS-700: Update the default values for kinesis stream name and staging bucket name.